refactor(core): move profile_types.ts to primtives
move profile_types.ts so the types can be used in Wiz code
