+++ packages/core/src/render3/after_render/manager.ts
@@ -18,7 +18,7 @@ import {type DestroyRef} from '../../linker/destroy_ref';
 import {NgZone} from '../../zone';
 import {AFTER_RENDER_SEQUENCES_TO_ADD, FLAGS, LView, LViewFlags} from '../interfaces/view';
 import {profiler} from '../profiler';
-import {ProfilerEvent} from '../profiler_types';
+import {ProfilerEvent} from '../../../primitives/profiler/src/profiler_types';
 import {markAncestorsForTraversal} from '../util/view_utils';
 import {AfterRenderPhase, AfterRenderRef} from './api';
