+++ packages/core/src/render3/debug/chrome_dev_tools_performance.ts
@@ -10,7 +10,7 @@ import {isTypeProvider} from '../../di/provider_collection';
 import {assertDefined, assertEqual} from '../../util/assert';
 import {performanceMarkFeature} from '../../util/performance';
 import {setProfiler} from '../profiler';
-import {Profiler, ProfilerEvent} from '../profiler_types';
+import {Profiler, ProfilerEvent} from '../../../primitives/profiler/src/profiler_types';
 import {stringifyForError} from '../util/stringify_utils';
 import {
   InjectorProfiler,
