+++ packages/core/src/render3/instructions/render.ts
@@ -22,7 +22,7 @@ import {
   TView,
 } from '../interfaces/view';
 import {profiler} from '../profiler';
-import {ProfilerEvent} from '../profiler_types';
+import {ProfilerEvent} from '../../../primitives/profiler/src/profiler_types';
 import {executeViewQueryFn, refreshContentQueries} from '../queries/query_execution';
 import {enterView, leaveView} from '../state';
 import {getComponentLViewByIndex, isCreationMode} from '../util/view_utils';
