+++ packages/core/src/render3/instructions/shared.ts
@@ -57,7 +57,7 @@ import {
 import {assertTNodeType} from '../node_assert';
 import {isNodeMatchingSelectorList} from '../node_selector_matcher';
 import {profiler} from '../profiler';
-import {ProfilerEvent} from '../profiler_types';
+import {ProfilerEvent} from '../../../primitives/profiler/src/profiler_types';
 import {
   getCurrentDirectiveIndex,
   getCurrentTNode,
